fix clean_up for data_generators; add feat_dir for jointdnn so that we don't need to prepare every time

+++ steps_tf/data_generator/frame_data_generator.py
@@ -104,7 +104,7 @@ class FrameDataGenerator:
 
 
   def __del__(self):
-    if self.clean_up:
+    if self.clean_up and os.path.exists(self.tmp_dir):
       shutil.rmtree(self.tmp_dir)
